Pulled generic AQHI handler logic out of controller.
